tests/wpt/tests/wasm/webapi/esm-integration/source-phase-blocked-by-csp.tentative.html

<!doctype html>
<title>Source phase imports blocked by CSP</title>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  setup({allow_uncaught_exception: true});

  const test_load = async_test(
      "Importing a WebAssembly module should be guarded by script-src CSP.");

  window.violationCount = 0;
  document.addEventListener("securitypolicyviolation", (e) => {
    test_load.step(() => {
      assert_equals("script-src-elem", e.violatedDirective);
    });
    if (++violationCount == 2) {
      test_load.done();
    }
  });
</script>
<script type=module src="./resources/cross-origin-wasm-static-source-phase-import.sub.js"></script>
<script type=module src="./resources/cross-origin-wasm-dynamic-source-phase-import.sub.js"></script>
